新增函数

+++ Serverfunc_Tan/src/main/java/com/toone/v3/platform/function/TanFunc.java
@@ -54,7 +54,7 @@ public class TanFunc implements IFunction {
             result = result.setScale(10, BigDecimal.ROUND_HALF_UP);
 
             outputVo.setSuccess(true);
-            outputVo.put(result.doubleValue());
+            outputVo.put(result);
         } catch (ServerFuncException e) {
             outputVo.setSuccess(false);
             outputVo.setMessage(e.getMessage());
